requirements.txt
amqp==2.5.1
billiard==3.6.1.0
blis==0.2.4
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
cymem==2.0.2
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==2.2.5
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-compat==1.0.15
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.1.0
django-hosts==3.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-storages-redux==1.3.3
django-stubs==1.1.0
django-timezone-field==3.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
environ==1.0
gunicorn==19.7.1
idna==2.6
importlib-metadata==0.20
kombu==4.6.4
lxml==4.3.4
markdown2==2.3.8
more-itertools==7.2.0
murmurhash==1.0.2
mypy==0.720
mypy-extensions==0.4.1
nltk==3.4.4
numpy==1.17.1
pandas==0.25.1
plac==0.9.6
preshed==2.0.1
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.6
requests==2.18.4
six==1.12.0
spacy==2.1.8
sqlparse==0.3.0
srsly==0.1.0
thinc==7.0.8
tqdm==4.35.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.22
vine==1.3.0
wasabi==0.2.2
whitenoise==4.1.2
zipp==0.6.0
